Bug 877751 - Handle Cell Broadcast System messages in the system app.
Restored previous functionality that shows CB messages recieved
on channel 50 to the lockscreen, all other messages are treated
as user alerts with the following behaviour:

If a message is received while in the lockscreen then a notification
is sent which wakes the screen, sounds a notification and displays
it on the lock screen, clicking on the notification will then show
a dialog with a full message.

If we are currently using the device the dialog will be displayed
immediately.