Bug 883341 - Ensure that the 'pick' activity is valid for Contacts
When the user tap on an <input type="file"> field, this triggers a
'pick' activity with no data, hence any app implementing the 'pick'
activity handler will be proposed. However, the Contacts app does not
know how to properly handle this case (in fact it does not make sense),
so we force the filtering to be done with the proper mime type: Contacts
app will not be proposed anymore for 'pick' activity triggered from
<input type="file"> user interaction. Thanks to Andrea Marchesini
(:baku) for suggesting this!