Bug 983459 - Notify of no voicemail number when tapping voicemail notification
If we don't have any voicemail number (on the SIM card nor configured in
Settings), when the user is tapping on the voicemail notification,
nothing will happen. This is confusing, so we implement a modal dialog
to notify of the lack of voicemail number configured, and we add a link
to the Settings app to configure it.